Move Rule to the lexer module and delete ast.rs.
Since ast.rs is now empty, there's no need to keep it around.
